requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
alembic==1.17.0
    # via mlflow
annotated-doc==0.0.3
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via starlette
asttokens==3.0.0
    # via stack-data
blinker==1.9.0
    # via flask
cachetools==6.2.1
    # via
    #   google-auth
    #   mlflow-skinny
    #   mlflow-tracing
certifi==2025.10.5
    # via requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.4
    # via requests
click==8.3.0
    # via
    #   flask
    #   mlflow-skinny
    #   nltk
    #   uvicorn
cloudpickle==3.1.1
    # via mlflow-skinny
comm==0.2.3
    # via ipykernel
contourpy==1.3.3
    # via matplotlib
cryptography==46.0.3
    # via mlflow
cycler==0.12.1
    # via matplotlib
databricks-sdk==0.70.0
    # via
    #   mlflow-skinny
    #   mlflow-tracing
debugpy==1.8.17
    # via ipykernel
decorator==5.2.1
    # via ipython
docker==7.1.0
    # via mlflow
executing==2.2.1
    # via stack-data
fastapi==0.120.0
    # via
    #   spam (pyproject.toml)
    #   mlflow-skinny
flask==3.1.2
    # via
    #   flask-cors
    #   mlflow
flask-cors==6.0.1
    # via mlflow
fonttools==4.60.1
    # via matplotlib
gitdb==4.0.12
    # via gitpython
gitpython==3.1.45
    # via mlflow-skinny
google-auth==2.41.1
    # via databricks-sdk
graphene==3.4.3
    # via mlflow
graphql-core==3.2.6
    # via
    #   graphene
    #   graphql-relay
graphql-relay==3.2.0
    # via graphene
greenlet==3.2.4
    # via sqlalchemy
gunicorn==23.0.0
    # via mlflow
h11==0.16.0
    # via uvicorn
idna==3.11
    # via
    #   anyio
    #   requests
importlib-metadata==8.7.0
    # via
    #   mlflow-skinny
    #   opentelemetry-api
ipykernel==7.0.1
    # via spam (pyproject.toml)
ipython==9.6.0
    # via ipykernel
ipython-pygments-lexers==1.1.1
    # via ipython
itsdangerous==2.2.0
    # via flask
jedi==0.19.2
    # via ipython
jinja2==3.1.6
    # via flask
joblib==1.5.2
    # via
    #   nltk
    #   scikit-learn
jupyter-client==8.6.3
    # via ipykernel
jupyter-core==5.9.1
    # via
    #   ipykernel
    #   jupyter-client
kiwisolver==1.4.9
    # via matplotlib
mako==1.3.10
    # via alembic
markupsafe==3.0.3
    # via
    #   flask
    #   jinja2
    #   mako
    #   werkzeug
matplotlib==3.10.7
    # via
    #   spam (pyproject.toml)
    #   mlflow
    #   seaborn
    #   wordcloud
matplotlib-inline==0.2.1
    # via
    #   ipykernel
    #   ipython
mlflow==3.5.1
    # via spam (pyproject.toml)
mlflow-skinny==3.5.1
    # via mlflow
mlflow-tracing==3.5.1
    # via mlflow
nest-asyncio==1.6.0
    # via ipykernel
nltk==3.9.2
    # via spam (pyproject.toml)
numpy==2.3.4
    # via
    #   contourpy
    #   matplotlib
    #   mlflow
    #   pandas
    #   scikit-learn
    #   scipy
    #   seaborn
    #   wordcloud
opentelemetry-api==1.38.0
    # via
    #   mlflow-skinny
    #   mlflow-tracing
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
opentelemetry-proto==1.38.0
    # via
    #   mlflow-skinny
    #   mlflow-tracing
opentelemetry-sdk==1.38.0
    # via
    #   mlflow-skinny
    #   mlflow-tracing
opentelemetry-semantic-conventions==0.59b0
    # via opentelemetry-sdk
packaging==25.0
    # via
    #   gunicorn
    #   ipykernel
    #   matplotlib
    #   mlflow-skinny
    #   mlflow-tracing
pandas==2.3.3
    # via
    #   spam (pyproject.toml)
    #   mlflow
    #   seaborn
parso==0.8.5
    # via jedi
pexpect==4.9.0
    # via ipython
pillow==12.0.0
    # via
    #   matplotlib
    #   wordcloud
platformdirs==4.5.0
    # via jupyter-core
prompt-toolkit==3.0.52
    # via ipython
protobuf==6.33.0
    # via
    #   databricks-sdk
    #   mlflow-skinny
    #   mlflow-tracing
    #   opentelemetry-proto
psutil==7.1.2
    # via ipykernel
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.3
    # via stack-data
pyarrow==21.0.0
    # via mlflow
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycparser==2.23
    # via cffi
pydantic==2.12.3
    # via
    #   fastapi
    #   mlflow-skinny
    #   mlflow-tracing
pydantic-core==2.41.4
    # via pydantic
pygments==2.19.2
    # via
    #   ipython
    #   ipython-pygments-lexers
pyparsing==3.2.5
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   graphene
    #   jupyter-client
    #   matplotlib
    #   pandas
python-dotenv==1.2.1
    # via mlflow-skinny
pytz==2025.2
    # via pandas
pyyaml==6.0.3
    # via mlflow-skinny
pyzmq==27.1.0
    # via
    #   ipykernel
    #   jupyter-client
regex==2025.10.23
    # via nltk
requests==2.32.5
    # via
    #   databricks-sdk
    #   docker
    #   mlflow-skinny
rsa==4.9.1
    # via google-auth
scikit-learn==1.7.2
    # via
    #   spam (pyproject.toml)
    #   mlflow
scipy==1.16.2
    # via
    #   mlflow
    #   scikit-learn
seaborn==0.13.2
    # via spam (pyproject.toml)
six==1.17.0
    # via python-dateutil
smmap==5.0.2
    # via gitdb
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.44
    # via
    #   alembic
    #   mlflow
sqlparse==0.5.3
    # via mlflow-skinny
stack-data==0.6.3
    # via ipython
starlette==0.48.0
    # via fastapi
threadpoolctl==3.6.0
    # via scikit-learn
tornado==6.5.2
    # via
    #   ipykernel
    #   jupyter-client
tqdm==4.67.1
    # via nltk
traitlets==5.14.3
    # via
    #   ipykernel
    #   ipython
    #   jupyter-client
    #   jupyter-core
    #   matplotlib-inline
typing-extensions==4.15.0
    # via
    #   alembic
    #   anyio
    #   fastapi
    #   graphene
    #   mlflow-skinny
    #   opentelemetry-api
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   starlette
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via
    #   docker
    #   requests
uvicorn==0.38.0
    # via mlflow-skinny
wcwidth==0.2.14
    # via prompt-toolkit
werkzeug==3.1.3
    # via
    #   flask
    #   flask-cors
wordcloud==1.9.4
    # via spam (pyproject.toml)
zipp==3.23.0
    # via importlib-metadata
